+++ contrib/start-crawlers-legacy.sh
@@ -14,6 +14,4 @@ if ! test -f .venv/bin/monocle; then
     ./.venv/bin/python3 setup.py install
 fi
 
-PORT=${1:-9876}
-
-exec ./.venv/bin/uwsgi --http ":${PORT}" --manage-script-name --mount /app=monocle.webapp:app
+exec ./.venv/bin/monocle --elastic-conn "${ELASTIC_CONN:-localhost:9200}" crawler --config "${CONFIG:-etc/config.yaml}"