+++ haskell/src/Monocle/Api/Config.hs
@@ -17,13 +17,13 @@ module Monocle.Api.Config where
 import qualified Data.ByteString as BS
 import Data.Either.Validation (Validation (Failure, Success))
 import qualified Data.Map as Map
-import Data.Time.Clock (UTCTime)
+import Data.Text (dropWhileEnd)
 import qualified Dhall
 import qualified Dhall.Core
 import qualified Dhall.Src
 import qualified Dhall.TH
 import qualified Dhall.YamlToDhall as Dhall
-import Relude
+import Monocle.Prelude
 import System.Directory (getModificationTime)
 
 -- | Generate Haskell Type from Dhall Type
@@ -50,6 +50,13 @@ Dhall.TH.makeHaskellTypes
           Dhall.TH.SingleConstructor "Config" "Config" "./dhall-monocle/Monocle/Config.dhall"
         ]
   )
+
+-- | Some useful lens
+crawlersApiKeyLens :: Lens' Index Text
+crawlersApiKeyLens =
+  lens
+    (fromMaybe "CRAWLERS_API_KEY" . crawlers_api_key)
+    (\index newKey -> index {crawlers_api_key = Just newKey})
 
 -- | Embed the expected configuration schema
 configurationSchema :: Dhall.Core.Expr Dhall.Src.Src Void
@@ -115,16 +122,23 @@ loadConfig configPath = do
   -- the first constructor that fit.
   configTS <- liftIO $ getModificationTime configPath
   expr <- liftIO $ Dhall.dhallFromYaml loadOpt =<< BS.readFile configPath
-  pure $ case Dhall.extract Dhall.auto expr of
-    Success config ->
-      let configWorkspaces = workspaces config
-       in ReloadableConfig {..}
+  case Dhall.extract Dhall.auto expr of
+    Success config -> do
+      configWorkspaces <- traverse resolveEnv (workspaces config)
+      pure $ ReloadableConfig {..}
     Failure err -> error $ "Invalid configuration: " <> show err
   where
     configType = Dhall.Core.pretty configurationSchema
     loadOpt = Dhall.defaultOptions $ Just configType
 
 data ReloadableConfig = ReloadableConfig {configTS :: UTCTime, configPath :: FilePath, configWorkspaces :: [Index]}
+
+getSecret :: MonadIO m => Text -> Maybe Text -> m Text
+getSecret def keyM =
+  toText . fromMaybe (error $ "Missing environment: " <> env)
+    <$> lookupEnv (toString env)
+  where
+    env = fromMaybe def keyM
 
 reloadConfig :: MonadIO m => IORef ReloadableConfig -> m [Index]
 reloadConfig configRef = liftIO $ do
@@ -182,17 +196,22 @@ getAliases index = maybe [] (fmap toTuple) (search_aliases index)
 
 getCrawlerProject :: Crawler -> [Text]
 getCrawlerProject Crawler {..} = case provider of
-  GitlabProvider Gitlab {..} -> fromMaybe [] gitlab_repositories
+  GitlabProvider Gitlab {..} ->
+    let addOrgPrefix repo = removeTrailingSlash gitlab_organization <> "/" <> repo
+     in addOrgPrefix <$> fromMaybe [] gitlab_repositories
   _anyOtherProvider -> []
 
-getCrawlerOrganization :: Crawler -> [Text]
+removeTrailingSlash :: Text -> Text
+removeTrailingSlash = dropWhileEnd (== '/')
+
+getCrawlerOrganization :: Crawler -> Maybe Text
 getCrawlerOrganization Crawler {..} = case provider of
-  GitlabProvider Gitlab {..} -> fromMaybe [] gitlab_organizations
-  _anyOtherProvider -> []
+  GitlabProvider Gitlab {..} -> Just gitlab_organization
+  _anyOtherProvider -> Nothing
 
 emptyTenant :: Text -> [Ident] -> Index
 emptyTenant name idents' =
-  let crawlers_api_key = ""
+  let crawlers_api_key = Nothing
       crawlers = []
       projects = Nothing
       idents = Just idents'
@@ -245,3 +264,6 @@ getIdentByAliasFromIdents alias idents' = case isMatched <$> idents' of
   where
     isMatched :: Ident -> Maybe Text
     isMatched Ident {..} = if alias `elem` aliases then Just ident else Nothing
+
+resolveEnv :: MonadIO m => Index -> m Index
+resolveEnv = liftIO . mapMOf crawlersApiKeyLens getEnv'